requirements.txt
flask>=3.0.0
PyMuPDF>=1.24.0
pandas>=2.0.0
pdfplumber>=0.10.0
pywin32>=306
requests>=2.0.0
sqlalchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0